Some cleaning
