+++ src/Sidekick/Views/Settings/SettingsView.xaml
@@ -1,4 +1,4 @@
-<base:BaseTitleWindow
+<base:BaseWindow
     x:Class="Sidekick.Views.Settings.SettingsView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -118,4 +118,4 @@
                 Content="{loc:Loc Ok}" />
         </WrapPanel>
     </Grid>
-</base:BaseTitleWindow>
+</base:BaseWindow>
